+++ jolt-core/src/jolt/instruction/blt.rs
@@ -21,11 +21,11 @@ mod test {
     for _ in 0..256 {
       let x = rng.next_u64() as i64;
       let y = rng.next_u64() as i64;
-      jolt_instruction_test!(BLTInstruction(x, y), (x < y).into());
+      jolt_instruction_test!(BLTInstruction(x as u64, y as u64), (x < y).into());
     }
     for _ in 0..256 {
       let x = rng.next_u64() as i64;
-      jolt_instruction_test!(BLTInstruction(x, x), Fr::zero());
+      jolt_instruction_test!(BLTInstruction(x as u64, x as u64), Fr::zero());
     }
   }
 }
